requirements.txt
# Main requirements
streamlit>=1.32.0
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
pydantic>=1.10.0
python-dotenv>=1.0.0

# RAG system requirements
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0

# API and server requirements
fastapi
uvicorn

# Audio processing
pydub>=0.25.1
elevenlabs  # For advanced text-to-speech conversion
speechrecognition  # For speech-to-text functionality
# Note: simpleaudio is intentionally excluded as it requires system dependencies

# Utility libraries
requests>=2.31.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles

# Multi-LLM support
aisuite
anthropic
google-generativeai
